moved name input to own component
